+++ CHANGELOG.rst
@@ -42,7 +42,7 @@ Changed
 - Added generic init tests to base tests for transformers that take two columns as an input.
 - Refactored EqualityChecker tests in new format.
 - Bugfix to MeanResponseTransformer to ignore unobserved categorical levels
-
+- Added test_BaseTwoColumnTransformer base class for columns that require a list of two columns for input
 
 Removed
 ^^^^^^^